Code cleanup
Refactor the code and normalise the API

The helper functions are now factored into the body
of the publicly accessible functions.

The two ways to calculate contacs are clearly delineated
and the code that allows for the calculation of distances
only for the interacting chains has been rewritten to make
use of a named tuple.

Calculate atomic distances based on CMD argument