Change the format of the chain specification
from now on the first set of chains to be specified
_always_ refers to the protein and the second to the
ligand.

The ligand needs to be identified by chain and residue
name.

The contacts calculated from the ligand can now be
_intra_ as well as inter chain.

The C++ contact calculation is no longer part of the
code. The main reason for that it does not offer a
significant speed boost for protein-small molecule
interfaces and that in its current form it only
calculates inter chain distances.